adding search bar

+++ ui/app/controllers/passwords/edit.js
@@ -16,12 +16,14 @@ export default Controller.extend({
     },
     deletePassword(password) {
       console.log("/controllers/passwords/edit.js");
-  
-      this.store.findRecord('password', password.id, { backgroundReload: false }).then(function(record) {
-        record.destroyRecord(); // => DELETE to '/passwords/id'
-      });
-  
-      this.get('router').transitionTo('/passwords')    
+      
+      if (confirm("Are you sure?")) {
+        this.store.findRecord('password', password.id, { backgroundReload: false }).then(function(record) {
+          record.destroyRecord(); // => DELETE to '/passwords/id'
+        });
+    
+        this.get('router').transitionTo('/passwords')    
+      }
     }  
   },
 });
